services/ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
tensorflow-cpu==2.16.1
numpy==1.26.4
h5py==3.11.0
